# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/enrollments": {
      "post": {
        "operationId": "EnrollmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnrollmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Enrollment data is invalid"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Agent assignment is not permitted"
          },
          "404": {
            "description": "Patient, companion, or agent not found"
          },
          "409": {
            "description": "Patient is already enrolled or enrollment data conflicts"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a patient enrollment",
        "tags": [
          "enrollments"
        ]
      }
    },
    "/enrollments/patient/{patientId}": {
      "get": {
        "operationId": "EnrollmentsController_findAll",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnrollmentResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List a patient enrollments",
        "tags": [
          "enrollments"
        ]
      }
    },
    "/follow-ups": {
      "post": {
        "operationId": "FollowUpsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFollowUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Required agent assignment is missing"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Agents can only assign follow-ups to themselves"
          },
          "404": {
            "description": "Patient or agent not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a follow-up",
        "tags": [
          "follow-ups"
        ]
      }
    },
    "/follow-ups/{id}": {
      "get": {
        "operationId": "FollowUpsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "The authenticated user has no agent profile"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Agents can only access their own follow-ups"
          },
          "404": {
            "description": "Follow-up not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a follow-up",
        "tags": [
          "follow-ups"
        ]
      },
      "patch": {
        "operationId": "FollowUpsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFollowUpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "The authenticated user has no agent profile"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Agents can only update their own follow-ups"
          },
          "404": {
            "description": "Follow-up not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a follow-up",
        "tags": [
          "follow-ups"
        ]
      }
    },
    "/follow-ups/{id}/schedule-next": {
      "post": {
        "operationId": "FollowUpsController_next",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFollowUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Required agent assignment is missing"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Agents can only access their own follow-ups"
          },
          "404": {
            "description": "Follow-up, patient, or agent not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Schedule the next follow-up",
        "tags": [
          "follow-ups"
        ]
      }
    },
    "/follow-ups/{id}/reminders": {
      "post": {
        "operationId": "FollowUpsController_reminder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReminderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Required agent assignment is missing"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Agents can only access their own follow-ups"
          },
          "404": {
            "description": "Follow-up or agent not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a reminder from a follow-up",
        "tags": [
          "follow-ups"
        ]
      }
    },
...
    "/patients/{patientId}/diagnoses": {
      "post": {
        "operationId": "PatientDiagnosesController_create",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientDiagnosisDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDiagnosisResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload"
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          },
          "404": {
            "description": "Patient or follow-up not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record a patient diagnosis",
        "tags": [
          "Patient diagnoses"
        ]
      },
      "get": {
        "operationId": "PatientDiagnosesController_findAll",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientDiagnosisResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List a patient diagnosis history",
        "tags": [
          "Patient diagnoses"
        ]
      }
    },
...
  },
  "info": {
    "title": "FPC Backend API",
    "description": "FPC backend API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "EnrollmentFollowUpDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "WHATSAPP",
              "CALL",
              "VIDEO_CALL",
              "EMAIL",
              "IN_PERSON",
              "FACEBOOK"
            ]
          },
          "agentId": {
            "type": "string",
            "format": "uuid"
          },
          "notes": {
            "type": "string"
          },
          "scheduledAt": {
            "type": "string"
          },
          "completedAt": {
            "type": "string"
          }
        },
        "required": [
          "type"
        ]
      },
...
      "CreateEnrollmentDto": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "patient": {
            "$ref": "#/components/schemas/CreatePatientDto"
          },
          "followUp": {
            "$ref": "#/components/schemas/EnrollmentFollowUpDto"
          },
          "affiliationType": {
            "type": "string",
            "enum": [
              "SELF",
              "FAMILY_FRIEND"
            ]
          },
          "companionId": {
            "type": "string",
            "format": "uuid"
          },
          "companion": {
            "$ref": "#/components/schemas/CreateCompanionDto"
          },
          "details": {
            "$ref": "#/components/schemas/UpsertPatientDetailsDto"
          },
          "insurance": {
            "$ref": "#/components/schemas/EnrollmentInsuranceDto"
          },
          "sisAffiliation": {
            "$ref": "#/components/schemas/EnrollmentSisAffiliationDto"
          },
          "diagnosis": {
            "$ref": "#/components/schemas/EnrollmentDiagnosisDto"
          },
          "treatment": {
            "$ref": "#/components/schemas/EnrollmentTreatmentDto"
          },
          "medicalAppointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnrollmentMedicalAppointmentDto"
            }
          },
          "symptomReport": {
            "$ref": "#/components/schemas/EnrollmentSymptomReportDto"
          },
          "currentlyAttendingConsultations": {
            "type": "boolean"
          },
          "currentlyReceivingTreatment": {
            "type": "boolean"
          },
          "entrySource": {
            "type": "string",
            "maxLength": 50
          },
          "entrySubSource": {
            "type": "string",
            "maxLength": 50
          },
          "consentToContact": {
            "type": "boolean"
          },
          "consentToShareData": {
            "type": "boolean"
          },
          "requiresTransportation": {
            "type": "boolean"
          },
          "hasMobilityIssues": {
            "type": "boolean"
          },
          "isOncologicalPatient": {
            "type": "boolean"
          },
          "surveyAccepted": {
            "type": "boolean"
          },
          "followUpQualityRating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          }
        },
        "required": [
          "followUp",
          "affiliationType"
        ]
      },
      "EnrollmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "followUpId": {
            "type": "string",
            "format": "uuid"
          },
          "affiliationType": {
            "enum": [
              "SELF",
              "FAMILY_FRIEND"
            ],
            "type": "string"
          },
          "companionId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "currentlyAttendingConsultations": {
            "type": "boolean",
            "nullable": true
          },
          "currentlyReceivingTreatment": {
            "type": "boolean",
            "nullable": true
          },
          "entrySource": {
            "type": "string",
            "nullable": true
          },
          "entrySubSource": {
            "type": "string",
            "nullable": true
          },
          "consentToContact": {
            "type": "boolean",
            "nullable": true
          },
          "consentToShareData": {
            "type": "boolean",
            "nullable": true
          },
          "requiresTransportation": {
            "type": "boolean",
            "nullable": true
          },
          "hasMobilityIssues": {
            "type": "boolean",
            "nullable": true
          },
          "isOncologicalPatient": {
            "type": "boolean"
          },
          "surveyAccepted": {
            "type": "boolean"
          },
          "followUpQualityRating": {
            "type": "number",
            "nullable": true,
            "minimum": 1,
            "maximum": 5
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "patientId",
          "followUpId",
          "affiliationType",
          "companionId",
          "currentlyAttendingConsultations",
          "currentlyReceivingTreatment",
          "entrySource",
          "entrySubSource",
          "consentToContact",
          "consentToShareData",
          "requiresTransportation",
          "hasMobilityIssues",
          "isOncologicalPatient",
          "surveyAccepted",
          "followUpQualityRating",
          "createdAt"
        ]
      },
      "CreateFollowUpDto": {
        "type": "object",
        "properties": {
          "subjectPatientId": {
            "type": "string",
            "format": "uuid"
          },
          "interlocutorId": {
            "type": "string",
            "format": "uuid"
          },
          "agentId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "enum": [
              "WHATSAPP",
              "CALL",
              "VIDEO_CALL",
              "EMAIL",
              "IN_PERSON",
              "FACEBOOK"
            ],
            "type": "string"
          },
          "purpose": {
            "enum": [
              "FIRST_CONTACT",
              "ENROLLMENT",
              "FOLLOW_UP",
              "PSYCHOONCOLOGY_REFERRAL",
              "OTHER"
            ],
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "scheduledAt": {
            "type": "string"
          },
          "completedAt": {
            "type": "string"
          }
        },
        "required": [
          "subjectPatientId",
          "interlocutorId",
          "type",
          "purpose"
        ]
      },
      "FollowUpResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subjectPatientId": {
            "type": "string",
            "format": "uuid"
          },
          "interlocutorId": {
            "type": "string",
            "format": "uuid"
          },
          "agentId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "enum": [
              "WHATSAPP",
              "CALL",
              "VIDEO_CALL",
              "EMAIL",
              "IN_PERSON",
              "FACEBOOK"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "SCHEDULED",
              "COMPLETED",
              "CANCELLED",
              "NO_ANSWER"
            ],
            "type": "string"
          },
          "purpose": {
            "enum": [
              "FIRST_CONTACT",
              "ENROLLMENT",
              "FOLLOW_UP",
              "PSYCHOONCOLOGY_REFERRAL",
              "OTHER"
            ],
            "type": "string"
          },
          "scheduledAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "completedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "nextFollowUpId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "subjectPatientId",
          "interlocutorId",
          "agentId",
          "type",
          "status",
          "purpose",
          "scheduledAt",
          "completedAt",
          "notes",
          "nextFollowUpId",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateFollowUpDto": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "COMPLETED",
              "CANCELLED",
              "NO_ANSWER"
            ]
          },
          "completedAt": {
            "type": "string"
          }
        }
      },
      "CreateReminderDto": {
        "type": "object",
        "properties": {
          "subjectPatientId": {
            "type": "string",
            "format": "uuid"
          },
          "dueAt": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "assignedAgentId": {
            "type": "string",
            "format": "uuid"
          },
          "createdFromFollowUpId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "subjectPatientId",
          "dueAt",
          "description"
        ]
      },
      "ReminderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subjectPatientId": {
            "type": "string",
            "format": "uuid"
          },
          "createdFromFollowUpId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "assignedAgentId": {
            "type": "string",
            "format": "uuid"
          },
          "dueAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "enum": [
              "PENDING",
              "DONE",
              "DISMISSED"
            ],
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "resultingFollowUpId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "subjectPatientId",
          "createdFromFollowUpId",
          "assignedAgentId",
          "dueAt",
          "description",
          "status",
          "completedAt",
          "resultingFollowUpId",
          "createdAt"
        ]
      },
...
      "CreatePatientDiagnosisDto": {
        "type": "object",
        "properties": {
          "followUpId": {
            "type": "string",
            "format": "uuid"
          },
          "diagnosis": {
            "type": "string"
          },
          "cancerStage": {
            "type": "string",
            "enum": [
              "STAGE_1",
              "STAGE_2",
              "STAGE_3",
              "STAGE_4",
              "UNKNOWN"
            ]
          },
          "diagnosisDate": {
            "type": "string"
          },
          "healthCenterId": {
            "type": "string",
            "format": "uuid"
          },
          "diagnosisSpecialty": {
            "type": "string"
          },
          "symptomLeadingToCheckup": {
            "type": "string"
          },
          "waitTimeForDiagnosis": {
            "type": "string"
          },
          "hasMedicalReport": {
            "type": "boolean"
          },
          "changeReason": {
            "type": "string"
          }
        },
        "required": [
          "followUpId",
          "diagnosis"
        ]
      },
...
    }
  }
}